Move data portion of `HasRelationMetadata` to dbt/artifacts
