+++ core/dbt/contracts/graph/components.py
@@ -2,10 +2,9 @@ from dataclasses import dataclass, field
 from datetime import timedelta
 from dbt.artifacts.resources import (
     FreshnessThreshold as FreshnessThresholdResource,
+    HasRelationMetadata as HasRelationMetadataResource,
     Time as TimeResource,
 )
-from dbt_common.contracts.util import Replaceable
-from dbt_common.dataclass_schema import dbtClassMixin
 from typing import Dict, Optional
 
 
@@ -45,20 +44,7 @@ class FreshnessThreshold(FreshnessThresholdResource):
 
 # Metrics, exposures,
 @dataclass
-class HasRelationMetadata(dbtClassMixin, Replaceable):
-    database: Optional[str]
-    schema: str
-
-    # Can't set database to None like it ought to be
-    # because it messes up the subclasses and default parameters
-    # so hack it here
-    @classmethod
-    def __pre_deserialize__(cls, data):
-        data = super().__pre_deserialize__(data)
-        if "database" not in data:
-            data["database"] = None
-        return data
-
+class HasRelationMetadata(HasRelationMetadataResource):
     @property
     def quoting_dict(self) -> Dict[str, bool]:
         if hasattr(self, "quoting"):
